pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "prd-agent"
version = "1.0.0"
description = "AI-powered PRD Generation Agent for Legacy Code Migration using LangChain, Temporal, and Qdrant"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "PRD Agent Team", email = "team@prdagent.dev" }
]
keywords = ["prd", "langchain", "temporal", "qdrant", "migration", "ai-agents"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    # Core LangChain
    "langchain>=0.3.14",
    "langchain-openai>=0.3.0",
    "langchain-community>=0.3.14",
    "langchain-qdrant>=0.2.0",
    "langchain-text-splitters>=0.3.4",
    
    # Vector Store
    "qdrant-client>=1.12.1",
    
    # Temporal Workflow
    "temporalio>=1.9.0",
    
    # OpenAI
    "openai>=1.59.5",
    
    # Minio Storage
    "minio>=7.2.12",
    
    # Atlassian Integration
    "atlassian-python-api>=3.41.16",
    
    # Image Processing
    "pillow>=11.1.0",
    "pytesseract>=0.3.13",
    
    # Document Processing
    "python-docx>=1.1.2",
    "pypdf>=5.1.0",
    "markdown>=3.7",
    
    # Data Processing
    "pydantic>=2.10.4",
    "pydantic-settings>=2.7.0",
    
    # CLI
    "typer>=0.15.1",
    "rich>=13.9.4",
    
    # Utilities
    "python-dotenv>=1.0.1",
    "aiohttp>=3.11.11",
    "httpx>=0.28.1",
    "structlog>=24.4.0",
    "tenacity>=9.0.0",
    
    # Code Analysis
    "tree-sitter>=0.23.4",
    "javalang>=0.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
    "pytest-cov>=6.0.0",
    "black>=24.10.0",
    "ruff>=0.8.6",
    "mypy>=1.14.1",
    "pre-commit>=4.0.1",
]

[project.scripts]
prd-agent = "src.cli.commands:app"
prd-worker = "src.worker.temporal_worker:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py311', 'py312']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=term-missing"

